Prisoners Dilemma
Prisoners Dilemma. In this case, the counts correspond to the outcomes of the game. The states |00> and |01> can be interpreted as both players cooperating and player A cooperating while player B defects, respectively. The states |10> and |11> can be interpreted as player A defecting while player B cooperates and both players defecting, respectively.