Bit Flip Error Correction
Bit flip error correction using the simplest error correction code - the three-qubit bit flip code.

In the bit flip code, we encode a logical 0 as 000 and a logical 1 as 111 using three physical qubits. If a bit flip error happens in one of the qubits, we can still determine the original logical value by majority vote.