Shannon Entropy Of A Black Hole - Quhole
This one is cool. The Shannon entropy of a system is a measure of the uncertainty or randomness associated with a set of probabilities. Here the entropy provides a measure of how information is spread across the states. This initializes the first two qubits in a superposition, applies a controlled-not (CNOT) operation to entangle them with the next two qubits, and then applies random unitary transformations to the last two qubits.   

The measurement outcomes are over 4 qubits, so there are 2^4=16 possible outcomes.  

If the system were completely random, with each of the 16 outcomes equally likely, the entropy would be: S = - ∑ 16 i = 1/16 log_2  1/16 = log_2 16 = 4. (We got 3.959)  

The first two qubits are initialized into a superposition, and the information scrambling isn't entirely random, so you might expect the entropy to be somewhat less than 4. Here we got 3.959.