QSignature + SHA256 + EC
The input message is processed using the SHA-256 hash algorithm, like #Bitcoin, creating a 256 bit hash. The last 8 bits of this hash are then selected - this truncation simplifies the quantum encoding process as above.

A quantum circuit is prepared, using three physical qubits for each bit in the truncated hash. If there are 8 bits in the truncated hash, the circuit will consist of 24 physical qubits and 8 classical bits for measurement.

Each bit in the truncated hash is encoded in a quantum state using the 3 qubit bit-flip code. If the bit is 1, an X-gate (NOT gate) is applied to the corresponding qubit, and then two CNOT gates are used for error correction.

The bit-flip code helps in detecting and correcting errors in the quantum state. It can correct if a single qubit in the three qubit group has been flipped. This adds robustness against noise and errors in the quantum computation. 💪🏽

The quantum circuit is measured, collapsing the quantum state into classical bits. The measurement results are collected in an array, corresponding to the encoded bits, as you see on the screenshot below.

The result of 1024 measurements on 10001010 (in our histogram) is the most likely measurement outcome we would expect, corresponding to the encoded truncated hash of our original message. The result likely indicates our simulation was relatively noise free and the error correction worked as intended.