Quantum Entanglement in 3D Vector Space
In this experiment using 
@qiskit
, we construct a quantum circuit to generate an entangled state using two #qubits. Once achieved, we visualize the degree of this entanglement in a 3D scatter plot, first time I've visualized entanglement this way! Subsequently, we evolve the state using a random unitary transformation and then visualize the entanglement of the newly evolved state, providing a graphical representation of how entanglement can change under quantum operations. All entangled states are uniform here.

Quantum Circuit Initialization:
A #quantum circuit with two qubits is initialized. These qubits will be used to produce the #entangled state.

Entangled State Creation:
A Hadamard gate (qc.h(0)) is applied to the first qubit, creating a superposition. Following that, a controlled-X gate (http://qc.cx(0, 1)) creates an entanglement between the two qubits.

Statevector Retrieval:
We utilize Qiskit's Aer module to simulate the quantum circuit and retrieve the statevector that represents the quantum state of the circuit after entanglement.

Entanglement Degree Calculation:
A custom function, entanglement_degree, computes a degree of entanglement by taking the outer product of the statevector and its complex conjugate. The degree is essentially the sum of off-diagonal elements, giving us an abstract measure of the entanglement.

degree = ∑_(i≠j)*  | M_(ij) ​∣
Where M_(ij)​ are the elements of matrix M, and the sum runs over all off-diagonal elements.

3D Visualization Setup:
We set up a 3D grid using np.meshgrid. This will serve as the canvas for our entanglement visualization.

Entanglement Visualization:
The computed entanglement degree is visualized as color intensities on a 3D scatter plot. The colormap viridis provides a gradient from yellow (low) to purple (high), representing the degree of entanglement.

Time Evolution of the State:
A random unitary transformation is applied to the initial entangled state, effectively evolving the quantum state. This is to understand how the state (and its entanglement) can change under quantum operations.

Visualization of the Evolved State:
Just like the initial state, the evolved state's degree of entanglement is computed and then visualized on a separate 3D scatter plot.

Results:
The Cube from 0 to 1 on each axis is merely a representational canvas. Each point in this cube doesn't correspond to a particular quantum state or a part of it. Instead, the entire space is taken to show variation (if any) of our metric: the degree of entanglement. We chose a cube from 0 to 1 in each dimension for simplicity and normalization, meaning the values are bounded between 0 and 1. This helps in visualization.

Each dot represents a value of the degree of entanglement. In our case, since we are using the same #statevector throughout the entire cube, the entanglement degree doesn't vary spatially. So, every dot has the same degree of entanglement, and its color signifies this value.

The colormap viridis is used to visually depict the degree of entanglement. In this colormap, lower values are represented with a more purple color and higher values with a more yellow color. The transition from purple to yellow goes through light green to darker green. The color of each dot is a direct representation of the entanglement degree at that point. Since our entanglement degree doesn't vary spatially, every dot should have the same color, depicting the same entanglement degree.